refactor: 파일명 리팩토링 #48
기존 ts 카멜->파스칼케이스 변경
프리펩의 경우 .prefab.ts로 바꿈
(applePrefab.ts-->Apple.prefab.ts)
파일명 바뀌면서 따라오는 경로명 바꿈
